Don't clobber original definitions if shut-up is loaded twice
Previously, shut-up would save the original functions for message
and load whenever shut-up.el was loaded. This assumed that
shut-up.el was only loaded once, and would only be loaded when message
and load had their original values.

As a result, the following code would break:

    ;; This sets `shut-up-load-original' to the original `load' function,
    ;; and likewise for `shut-up-write-region-original'.
    (load "path/to/shut-up.el") ;; or (require 'shut-up)

    (shut-up
      ;; This sets `shut-up-load-original' to `shut-up-load'.
      (load "path/to/shut-up.el")

      ;; Since `load' has been bound to `shut-up-load', this infinitely
      ;; recurses ("Variable binding depth exceeds max-specpdl-size") when
      ;; we try to call the original with `shut-up-load-original'.
      (load "path/to/shut-up.el"))

This causes problems with Cask (see cask/cask#367), which uses shut-up
when byte-compiling files.

Cask installs its own copy of shut-up (see the variable
cask-bootstrap-packages in cask-bootstrap.el), making this change
difficult to test.

If Cask installs a package that depends on shut-up (such as
undercover), it ends up byte-compiling shut-up.el inside a
(shut-up ...) form. This only occurs when cask is not run with
--verbose, so that was used as a workaround.

As of Emacs 26.1 (specifically emacs-mirror/emacs@50dce3c),
byte-compiled packages are reloaded, triggering this bug. Further
byte-compilation is broken. It's not clear to me why this doesn't
apply to all packages (e.g. undercover with tuareg is fine).

Ensure that we don't have side effects if shut-up is repeatedly
loaded.